integrates a jenkins variation with a java distribution from oracle
* docker-eula-java will be prepended to the default entrypoint

+++ .travis/build.sh
@@ -42,6 +42,8 @@ case "$CUSTOM" in
         docker run --rm "$TAG:$TAGSPECIFIER" -v
         ;;
     * )
+        if [ "$JAVA" = "8-jdk" ]; then sed -i -r '/ENTRYPOINT/ s!/bin/tini!docker-eula-java", "/bin/tini!g' "$PROJECT/Dockerfile"; fi
+
         patch -p1 --no-backup-if-mismatch --directory=$PROJECT < .patch/Dockerfile.patch
 
         mktini
@@ -51,6 +53,12 @@ case "$CUSTOM" in
                      --build-arg JENKINS_SHA="$SHA"         \
                      "$PROJECT"
 
-        docker run --rm "$TAG:$TAGSPECIFIER" java -jar /usr/share/jenkins/jenkins.war --version
+        if [ "$JAVA" = "8-jdk" ]; then
+             if docker run --rm -e eula-java=accept "$TAG:$TAGSPECIFIER" java -jar /usr/share/jenkins/jenkins.war --version; then             true; fi
+             if docker run --rm -e eula-java=       "$TAG:$TAGSPECIFIER" java -jar /usr/share/jenkins/jenkins.war --version; then false; else true; fi
+        else
+                docker run --rm                     "$TAG:$TAGSPECIFIER" java -jar /usr/share/jenkins/jenkins.war --version
+        fi
+
         ;;
 esac
